Split out Obsolete trigger version of AdaptiveCostEffect updated to new version:   [DAL_716] Vendetta   [GIL_518] Wing Blast
