+++ SabberStoneCore/src/CardSets/Standard/TrollCardsGen.cs
@@ -746,7 +746,7 @@ namespace SabberStoneCore.CardSets.Standard
 			// --------------------------------------------------------
 			cards.Add("TRL_313", new Power {
 				PowerTask = new DamageTask(4, EntityType.TARGET),
-				Aura = new AdaptiveCostEffect(1, TriggerType.TURN_START, triggerCondition: SelfCondition.ElementalPlayedLastTurn)
+				Aura = new AdaptiveCostEffect(p=>1, EffectOperator.SET, triggerCondition: SelfCondition.ElementalPlayedLastTurn)
 			});
 
 			// ------------------------------------------- SPELL - MAGE
@@ -821,7 +821,7 @@ namespace SabberStoneCore.CardSets.Standard
 			// - RUSH = 1
 			// --------------------------------------------------------
 			cards.Add("TRL_300", new Power {
-				Aura = new AdaptiveCostEffect(
+				Aura = new AdaptiveCostEffectObsolete(
 					initialisationFunction: p =>
 					{
 						int sum = 0;
@@ -1098,7 +1098,7 @@ namespace SabberStoneCore.CardSets.Standard
 			cards.Add("TRL_408", new Power {
 				// TODO: Possible performance overhead
 				//Aura = new AdaptiveCostEffect(p => p.Controller.PlayHistory.Count(h => h.SourceCard.Type == CardType.SPELL))
-				Aura = new AdaptiveCostEffect(
+				Aura = new AdaptiveCostEffectObsolete(
 					initialisationFunction: p => p.Card.Cost - p.Controller.PlayHistory.Count(h => h.SourceCard.Type == CardType.SPELL),
 					triggerValueFunction: p => -1,
 					trigger: TriggerType.AFTER_CAST,
