+++ SabberStoneCoreTest/src/Basic/UpdatedMechanics.cs
@@ -143,7 +143,7 @@ namespace SabberStoneCoreTest.Basic
 			for (int i = 1; i < 3; i++)
 			{
 				IPlayable copied = game.CurrentPlayer.HandZone[game.CurrentPlayer.HandZone.Count - i];
-				if (!(copied.OngoingEffect is AdaptiveCostEffect))
+				if (!(copied.OngoingEffect is IAdaptiveCostEffect))
 					Assert.Equal(1, copied.Cost);
 
 				if (copied is Minion m)
